+++ packages/better-effect-mq-mysql/src/layer.ts
@@ -1,1 +1,2 @@
 export { MySqlJobStore } from './MySqlJobStore'
+export { MySqlJobScheduleStore } from './schedule'